Split loader into module and cli
